Float-based heat Actual heat consumption Boilers and mixers consume heat and respect thermostat

+++ src/main/java/com/deltasf/createpropulsion/heat/HeatBuffer.java
@@ -4,17 +4,17 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraftforge.common.util.INBTSerializable;
 
 public class HeatBuffer implements IHeatSource, INBTSerializable<CompoundTag> {
-    protected int heat;
-    protected int capacity;
+    protected float heat;
+    protected float capacity;
 
-    public HeatBuffer(int initialHeat, int capacity) {
+    public HeatBuffer(float initialHeat, float capacity) {
         this.heat = initialHeat;
         this.capacity = capacity;
     }
 
     @Override
-    public int extractHeat(int amount, boolean simulate) {
-        int heatToExtract = Math.min(heat, amount);
+    public float extractHeat(float amount, boolean simulate) {
+        float heatToExtract = Math.min(heat, amount);
         if (!simulate) {
             heat -= heatToExtract;
         }
@@ -22,31 +22,31 @@ public class HeatBuffer implements IHeatSource, INBTSerializable<CompoundTag> {
     }
 
     @Override
-    public int getHeatStored() {
+    public float getHeatStored() {
         return heat;
     }
 
     @Override
-    public int getMaxHeatStored() {
+    public float getMaxHeatStored() {
         return capacity;
     }
 
     @Override
-    public void generateHeat(int amount) {
+    public void generateHeat(float amount) {
         heat = Math.min(capacity, heat + amount);
     }
 
     @Override
     public CompoundTag serializeNBT() {
         CompoundTag tag = new CompoundTag();
-        tag.putInt("Heat", this.heat);
-        tag.putInt("Capacity", this.capacity);
+        tag.putFloat("Heat", this.heat);
+        tag.putFloat("Capacity", this.capacity);
         return tag;
     }
 
     @Override
     public void deserializeNBT(CompoundTag nbt) {
-        this.heat = nbt.getInt("Heat");
-        this.capacity = nbt.getInt("Capacity");
+        this.heat = nbt.getFloat("Heat");
+        this.capacity = nbt.getFloat("Capacity");
     }
 }
